scripts/requirements.txt
supabase==2.13.2
python-dotenv==1.0.0
requests==2.31.0

